httpd_raw: fix detection of Content-Length header
HTTP headers are case-insensitiv, so using strnstr to find the Content-Length
header fails with browsers which use other case variants that the one defined
in the code.

To get this work correctly, we need a strncasestr function which might not
be available on some platforms. Thus, we need to define one ourself, upon
user request. Use existing code as much as possible for this.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>